+++ CHANGELOG.md
@@ -8,13 +8,14 @@ The format is inspired by *Keep a Changelog* while remaining focused on practica
 
 # Version 1.8 (Build 2620423) 2026-07-23
 
-## Bug
+## Resolved
 - sgnd-install overwrite the ownership of root, /etc, /usr, etcetera
-
-## Changed
 - sgnd-install.sh to never overwrite metadata of existing folders in teh target tree
 - prepare-release.sh to set ownership of repository folders to root:root
+
+## Changed
 - added toggle button to sgnd-console to switch between root and non-root
+- added left and right arrow for page navigation in sgnd-console
 
 # Version 1.8 (Build 2620413) 2026-07-23